Splits Plugin into Plugin.InputCustomizer and Plugin.OutputCustomizer to separate concerns.
